requirements.txt
asgiref==3.4.1
backoff==1.10.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Deprecated==1.2.12
Flask==2.0.1
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-trace==0.24.0
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
opentelemetry-api==1.3.0
opentelemetry-exporter-gcp-trace==1.0.0
opentelemetry-exporter-otlp-proto-grpc==1.3.0
opentelemetry-instrumentation==0.22b0
opentelemetry-instrumentation-flask==0.22b0
opentelemetry-instrumentation-requests==0.22b0
opentelemetry-instrumentation-wsgi==0.22b0
opentelemetry-propagator-gcp==1.0.0
opentelemetry-proto==1.3.0
opentelemetry-resourcedetector-gcp==1.0.0a0
opentelemetry-sdk==1.3.0
opentelemetry-semantic-conventions==0.22b0
opentelemetry-util-http==0.22b0
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
